refactor(config): extract ROM scanning + system detection out of SqliteConfigProvider
Phase 4 / H second slice. Move the ROM-scanning machinery and
system-detection/loading cluster (22 methods, ~900 lines) verbatim out
of SqliteConfigProvider into a new part file,
providers/sqlite_config_provider/scanning.dart, as
`extension SqliteConfigScanning on SqliteConfigProvider`.

Moved: scanSystems / _scanRomsInBackground / _scanSystemRoms /
refreshSystem / rescanSystem / rescanSystemSilent / refreshAllFilesAccess /
addRomFolder / removeRomFolder / selectRomFolder / clearConfig /
getQuickStats + the load/detect helpers (_loadConfig,
_loadAvailableSystems, reloadSystemDefinitions, _loadHiddenSystems,
toggleSystemHidden, _loadAvailableEmulators, _loadDetectedSystems,
refreshDetectedSystems, _refreshDetectedSystemsFromDatabase,
_sortDetectedSystems).

Same (a)-split kit as #166: public extension (methods are public API
bound by external importers); notifyListeners() routed through the host
`_notify()` bridge (24 calls); the static `_log` host-qualified as
`SqliteConfigProvider._log` (24 refs) — both required from an extension.
Both substitutions are provably equivalent.

Host retains the class decl, all state, lifecycle wiring
(initialize/_loadInitialData/reinitialize), and secondary-display logic.
Host 1507 -> 612; new part 931 lines. Host diff = one part directive +
pure deletions (remaining host byte-identical to main).
